chore(llmz): split unit tests from e2e tests
