+++ packages/llmz/tsconfig.json
@@ -14,5 +14,5 @@
     "typeRoots": ["node_modules/@types"]
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["**/*.test.ts", "src/**/*.test.ts", "**/__tests__", "node_modules"]
+  "exclude": ["node_modules"]
 }
