+++ packages/llmz/vitest.e2e.config.ts
@@ -1,0 +1,54 @@
+import 'dotenv/config'
+import fs from 'node:fs'
+import { defineConfig } from 'vitest/config'
+
+function textLoader() {
+  return {
+    name: 'text-loader',
+    transform(_src: string, id: string) {
+      if (id.endsWith('.txt')) {
+        const content = fs.readFileSync(id, 'utf-8')
+        return {
+          code: `export default ${JSON.stringify(content)};`,
+          map: null,
+        }
+      }
+    },
+  }
+}
+
+function mdLoader() {
+  return {
+    name: 'md-loader',
+    transform(_src: string, id: string) {
+      if (id.endsWith('.md')) {
+        const content = fs.readFileSync(id, 'utf-8')
+        return {
+          code: `export default ${JSON.stringify(content)};`,
+          map: null,
+        }
+      }
+    },
+  }
+}
+
+export default defineConfig({
+  plugins: [textLoader(), mdLoader()],
+  resolve: {
+    extensions: ['.js', '.ts', '.json', '.txt', '.md'],
+  },
+  assetsInclude: '**/*.md',
+  test: {
+    retry: 2, // because LLMs can fail
+    testTimeout: 60_000, // because LLMs can be slow
+    teardownTimeout: 10_000,
+    snapshotSerializers: ['./vitest.e2e.stack-trace-serializer.ts'],
+    snapshotEnvironment: './vitest.e2e.snapshot.ts',
+    maxConcurrency: 1,
+    isolate: false,
+    allowOnly: true,
+    pool: 'forks',
+    setupFiles: './vitest.e2e.setup.ts',
+    include: ['./e2e/**/*.test.ts'],
+  },
+})
